refactor: 레스토랑을 추가하는 함수 이름 변경
명확하게 하나의 음식점을 추가한다는 이름으로 변경.
